+++ inc/admin/acf.php
@@ -4,8 +4,8 @@
  *
  * @Author: Timi Wahalahti
  * @Date:   2020-01-10 16:11:23
- * @Last Modified by:   Timi Wahalahti
- * @Last Modified time: 2021-09-23 14:42:43
+ * @Last Modified by:   Elias Kautto
+ * @Last Modified time: 2021-11-18 13:53:58
  *
  * @package air-helper
  */
@@ -36,38 +36,17 @@ function air_helper_maybe_hide_acf() {
 } // end air_helper_maybe_hide_acf
 
 /**
- * If ACF Pro license key is defined in .env file, try to load the key from there
- * and activate ACF if not already activated or activation is corrupted.
+ * If ACF Pro license key is defined in .env file, try to load the key from there.
  *
- * Turn off by using `remove_action( 'admin_init', 'air_helper_get_acf_pro_license_from_env' )`
+ * Turn off by using `remove_action( 'admin_init', 'air_helper_define_acf_pro_license' )`
  *
- * @since 2.11.0
+ * @since 2.12.0
  */
-add_action( 'admin_init', 'air_helper_get_acf_pro_license_from_env' );
-function air_helper_get_acf_pro_license_from_env() {
-  // Bail if no ACF
-  if ( ! function_exists( 'acf_pro_get_license_key' ) ) {
-    return;
-  }
-
-  // Bail if no ACF
-  if ( ! method_exists( 'ACF_Admin_Updates', 'activate_pro_licence' ) ) {
-    return;
-  }
-
-  // Bail if no license key in .env
+add_action( 'admin_init', 'air_helper_define_acf_pro_license' );
+function air_helper_define_acf_pro_license() {
   if ( empty( getenv( 'ACF_PRO_KEY' ) ) ) {
     return;
   }
 
-  // Bail if license key is valid
-  if ( acf_pro_get_license_key() ) {
-    return;
-  }
-
-  // Force our license key into POST data, as ACF reads it from there
-  $_POST['acf_pro_licence'] = getenv( 'ACF_PRO_KEY' );
-
-  // Run ACF function to activate the license
-  ACF_Admin_Updates::activate_pro_licence();
-} // end air_helper_get_acf_pro_license_from_env
+  define( 'ACF_PRO_LICENSE', getenv( 'ACF_PRO_KEY' ) );
+} // end air_helper_define_acf_pro_license