+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Updated
+- Updated ACF Pro license key fetching, to use new ACF_PRO_LICENSE constant.
 
 ## [2.11.0] & [2.11.1] 2021-09-23
 ### Added